+++ libs/nest/prisma/migrations/20240724021629_add_effective_start_and_end_to_files/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "organization_files" ADD COLUMN     "effective_end_date" TIMESTAMP(3),
+ADD COLUMN     "effective_start_date" TIMESTAMP(3);